requirements.txt
numpy>=1.23
pandas>=1.5
scikit-learn>=1.3
matplotlib>=3.7
pyyaml>=6.0
tqdm>=4.66
datasets>=2.16
transformers>=4.40,<5
torch>=2.1,<3
accelerate>=0.26.0
